+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from "./algebra";
 export * from "./hashes";
+export * from "./hybrid-encryption";